Move GeoPolyActivity to geo module
