+++ collect_app/src/main/AndroidManifest.xml
@@ -143,8 +143,9 @@ the specific language governing permissions and limitations under the License.
             android:theme="@style/Theme.Collect"
             tools:replace="android:theme" />
         <activity
-            android:name=".activities.GeoPolyActivity"
-            android:theme="@style/Theme.Collect" />
+            android:name="org.odk.collect.geo.GeoPolyActivity"
+            android:theme="@style/Theme.Collect"
+            tools:replace="android:theme" />
         <activity android:name=".activities.FormMapActivity" />
         <activity android:name=".activities.BearingActivity" />
         <activity
