+++ geo/src/main/AndroidManifest.xml
@@ -10,5 +10,9 @@
         <activity
             android:name=".GeoPointMapActivity"
             android:theme="@style/Theme.MaterialComponents" />
+
+        <activity
+            android:name=".GeoPolyActivity"
+            android:theme="@style/Theme.MaterialComponents" />
     </application>
 </manifest>
