+++ geo/src/main/java/org/odk/collect/geo/DaggerSetup.kt
@@ -4,6 +4,7 @@ import dagger.Component
 import dagger.Module
 import dagger.Provides
 import org.odk.collect.geo.maps.MapFragmentFactory
+import org.odk.collect.location.tracker.LocationTracker
 import java.lang.UnsupportedOperationException
 import javax.inject.Singleton
 
@@ -15,6 +16,7 @@ interface GeoDependencyComponentProvider {
 @Singleton
 interface GeoDependencyComponent {
     fun inject(geoPointMapActivity: GeoPointMapActivity)
+    fun inject(geoPolyActivity: GeoPolyActivity)
 }
 
 @Module
@@ -27,6 +29,11 @@ open class GeoDependencyModule {
 
     @Provides
     open fun providesReferenceLayerSettingsNavigator(): ReferenceLayerSettingsNavigator {
+        throw UnsupportedOperationException("This should be overridden by dependent application")
+    }
+
+    @Provides
+    open fun providesLocationTracker(): LocationTracker {
         throw UnsupportedOperationException("This should be overridden by dependent application")
     }
 }
